slack_bot/requirements.txt
openai>=1.0.0
slack_sdk>=3.21.0
Flask>=2.2.0
requests>=2.32.0
python-dotenv>=1.0.0  # If you are loading environment variables from .env files
pillow>=11.2.0
numpy>=2.2.0
gunicorn
pycryptodome
